requirements.txt
flask==3.0.2
sentence-transformers==2.6.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
python-docx==0.8.11
PyPDF2==3.0.1
scikit-learn==1.4.1.post1
joblib==1.3.2
huggingface_hub==0.22.2
datasets==2.19.0
transformers==4.40.0
torch==2.2.2
gensim
scipy
google-generativeai